twitter_content_gen/requirements.txt
tweepy
transformers
torch
pandas
python-dotenv
numpy
scikit-learn